requirements.txt
flask
flask-smorest
flask-sqlalchemy
flask-jwt-extended
flask-migrate
passlib
python-dotenv
gunicorn
psycopg2
requests